+++ lib/src/render/filters/filters.dart
@@ -1,5 +1,7 @@
+export 'base_filter.dart';
 export 'blur_filter.dart';
 export 'color_filter.dart';
 export 'composer_filter.dart';
 export 'dropshadow_filter.dart';
 export 'glow_filter.dart';
+export 'outline_filter.dart';